[YS-559] refactor: 컴포넌트 관계에 따라 디렉토리 및 파일명 수정
