+++ src/app/home/components/Banner/Banner.tsx
@@ -1,6 +1,7 @@
 'use client';
 
 import Image from 'next/image';
+import { useState } from 'react';
 
 import {
   bannerCarousel,
@@ -19,6 +20,7 @@ import MobileBannerSecond from '@/assets/images/mobileBanner2.webp';
 import WebBanner from '@/assets/images/webBanner.png';
 import WebBannerSecond from '@/assets/images/webBanner2.png';
 import Icon from '@/components/Icon';
+import Spinner from '@/components/Spinner/Spinner';
 
 const BannerMap = [
   {
@@ -53,6 +55,8 @@ const Banner = () => {
     moveSlide,
   });
 
+  const [isLoading, setIsLoading] = useState(false);
+
   return (
     <div className={bannerLayout}>
       <div
@@ -71,15 +75,22 @@ const Banner = () => {
           {BannerMap.map((banner, idx) => (
             <picture key={idx} style={{ display: 'block', flex: '0 0 100%' }}>
               <source media="(max-width: 767px)" srcSet={banner.mobileSrc.src} />
-              <Image
-                key={idx}
-                src={banner.webSrc}
-                alt={banner.alt}
-                className={bannerImage}
-                priority
-                width={1000}
-                height={80}
-              />
+              {isLoading ? (
+                <Spinner />
+              ) : (
+                <Image
+                  key={idx}
+                  src={banner.webSrc}
+                  alt={banner.alt}
+                  className={bannerImage}
+                  priority
+                  width={1000}
+                  height={80}
+                  onLoad={() => {
+                    setIsLoading(false);
+                  }}
+                />
+              )}
             </picture>
           ))}
         </div>
