+++ src/app/home/components/ExperimentPostSection/ExperimentPostCardListContainer/ExperimentPostCardList/ExperimentPostCard/ExperimentPostCard.css.ts
