+++ src/app/home/components/ExperimentPostSection/ExperimentPostCardListContainer/ExperimentPostCardList/ExperimentPostCardList.css.ts
